live-coding-outputs/2025_04_21_Sydney/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
sqlmodel>=0.0.16
aiosqlite>=0.19.0
python-dotenv>=1.0.0

# Testing dependencies
pytest>=8.0.0
pytest-asyncio>=0.23.0 # For running async tests
httpx>=0.27.0 # For making async requests in tests
